% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_docx.R
\name{output_docx}
\alias{output_docx}
\title{output_docx}
\usage{
output_docx(..., gap = " ", filename = "output.docx",
  captions = NULL, digits = 2)
}
\arguments{
\item{...}{One or more data.frames}

\item{gap}{Separator between each table, the default is a SPACE}

\item{filename}{File name of the .docx}

\item{captions}{Title for each table}

\item{digits}{Number of digits shown in the table, the default is 2}
}
\description{
A simple function that helps to output tables in .docx
}
\examples{

output_docx(cars,iris)


}
